+++ monitor/ui/config.js
@@ -1,0 +1,3 @@
+window.MONITOR_CONFIG = {
+  apiUrl: 'http://127.0.0.1:4310'
+};